requirements.txt
Flask~=3.1.1
twilio~=9.6.1
sentence-transformers~=4.1.0
numpy~=2.0.2
python-dotenv~=1.1.0
faiss-cpu~=1.11.0
supabase~=2.15.2
pytest~=8.3.5
qdrant-client~=1.14.2